Move create_example_plots to examples/utils.py
