+++ examples/Backtesting/custom_blackbox.py
@@ -18,7 +18,7 @@ from baybe.recommenders import RandomRecommender
 from baybe.searchspace import SearchSpace
 from baybe.simulation import simulate_scenarios
 from baybe.targets import NumericalTarget
-from baybe.utils.plotting import create_example_plots
+from examples.utils import create_example_plots
 
 ### Settings
