+++ examples/Backtesting/full_lookup.py
@@ -24,7 +24,7 @@ from baybe.recommenders import RandomRecommender
 from baybe.searchspace import SearchSpace
 from baybe.simulation import simulate_scenarios
 from baybe.targets import NumericalTarget
-from baybe.utils.plotting import create_example_plots
+from examples.utils import create_example_plots
 
 ### Parameters for a full simulation loop
