+++ examples/Transfer_Learning/backtesting.py
@@ -24,7 +24,7 @@ from baybe.searchspace import SearchSpace
 from baybe.simulation import simulate_scenarios, simulate_transfer_learning
 from baybe.targets import NumericalTarget
 from baybe.utils.dataframe import arrays_to_dataframes
-from baybe.utils.plotting import create_example_plots
+from examples.utils import create_example_plots
 
 ### Settings
